backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
argcomplete==3.1.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
blobfile==3.0.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
eval_type_backport==0.2.2
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.29.3
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
llama_stack==0.1.8
llama_stack_client==0.1.8
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.2.0
mypy==1.5.1
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.78.0
overrides==7.7.0
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pipx==1.2.0
platformdirs==4.2.2
pluggy==1.3.0
proglog==0.1.12
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==25.1.0
pyarrow==19.0.1
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.11.0
pydantic_core==2.33.0
pydub==0.25.1
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
pytest==7.4.2
pytest-timeout==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.18.1
scikit-learn==1.6.1
scipy==1.15.1
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
tabulate==0.9.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.3.0
together==1.5.3
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
userpath==1.9.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.11
yarl==1.18.3
